+++ lib/chromic_pdf/pdf/protocols/navigate.ex
@@ -12,6 +12,17 @@ defmodule ChromicPDF.Navigate do
       await_response(:cookie_set, [])
     end
 
+    if_option :timezone do
+      call(
+        :timezone,
+        "Emulation.setTimezoneOverride",
+        &%{"timezoneId" => Map.fetch!(&1, :timezone)},
+        %{}
+      )
+
+      await_response(:timezone_set, [])
+    end
+
     if_option {:source_type, :html} do
       call(:get_frame_tree, "Page.getFrameTree", [], %{})
       await_response(:frame_tree, [{["frameTree", "frame", "id"], "frameId"}])